+++ src/main/java/uz/dckroff/statisfy/model/NotificationSettings.java
@@ -1,11 +1,11 @@
 package uz.dckroff.statisfy.model;
 
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
 import java.time.LocalTime;
 
 @Entity